backend/requirements.txt
## `backend/requirements.txt`
# ─────────────────────────────────────────────────────────────────────
# This file lists the Python dependencies for the FastAPI backend.
fastapi>=0.115.1           # latest w/ Python 3.13 wheel
httpx[http2]>=0.28
uvicorn[standard]>=0.34
sqlmodel>=0.0.24
asyncpg>=0.30
redis>=5.0.1
aioredis>=2.0.1
aiosqlite>=0.21
python-dotenv>=1.1
python-json-logger>=2.0
loguru>=0.7
tldextract>=5.1
gunicorn>=21.2